# perf(user): optimize getAchievementsEarnedBetween()