--- a/app/Helpers/render/user.php
+++ b/app/Helpers/render/user.php
@@ -71,25 +71,6 @@ function renderUserCard(string|array $user): string
     ]);
 }
 
-function getCompletedAndIncompletedSetsCounts(array $userCompletedGamesList): array
-{
-    $completedSetsCount = 0;
-    $incompletedSetsCount = 0;
-
-    foreach ($userCompletedGamesList as $game) {
-        $nextMaxPossible = $game['MaxPossible'];
-        $nextNumAwarded = $game['NumAwarded'];
-
-        if ($nextNumAwarded == $nextMaxPossible) {
-            $completedSetsCount++;
-        } else {
-            $incompletedSetsCount++;
-        }
-    }
-
-    return ['completedSetsCount' => $completedSetsCount, 'incompletedSetsCount' => $incompletedSetsCount];
-}
-
 function RenderCompletedGamesList(
     array $userCompletedGamesList,
     string $username,
@@ -100,28 +81,25 @@ function RenderCompletedGamesList(
     echo "<h3>Completion Progress</h3>";
 
     $checkedAttribute = $isInitiallyHidingCompletedSets ? 'checked' : '';
-    $setsCounts = getCompletedAndIncompletedSetsCounts($userCompletedGamesList);
-    if ($setsCounts['completedSetsCount'] > 0 && $setsCounts['incompletedSetsCount'] > 0) {
-        echo "<div class='flex w-full items-center justify-between mb-2'>";
-        echo <<<HTML
-            <label class="flex items-center gap-x-1">
-                <input 
-                    type="checkbox" 
-                    id="hide-user-completed-sets-checkbox" 
-                    onchange="toggleUserCompletedSetsVisibility()"
-                    $checkedAttribute
-                >
-                    Hide completed games
-                </input>
-            </label>
-        HTML;
 
-        if (config('feature.beat')) {
-            echo "<a href='" . route('user.completion-progress', $username) . "'>See more...</a>";
-        }
+    echo "<div class='flex w-full items-center justify-between mb-2'>";
+    echo <<<HTML
+        <label class="flex items-center gap-x-1">
+            <input 
+                type="checkbox" 
+                id="hide-user-completed-sets-checkbox" 
+                onchange="toggleUserCompletedSetsVisibility()"
+                $checkedAttribute
+            >
+                Hide completed games
+            </input>
+        </label>
+    HTML;
 
-        echo "</div>";
+    if (config('feature.beat')) {
+        echo "<a href='" . route('user.completion-progress', $username) . "'>See more...</a>";
     }
+    echo "</div>";
 
     echo "<div id='usercompletedgamescomponent'>";