--- a/database/migrations/platform/2023_10_30_000000_update_player_achievements_table.php
+++ b/database/migrations/platform/2023_10_30_000000_update_player_achievements_table.php
@@ -10,14 +10,17 @@ return new class() extends Migration {
     public function up(): void
     {
         Schema::table('player_achievements', function (Blueprint $table) {
-            $table->index(['user_id', 'unlocked_at', 'unlocked_hardcore_at', 'achievement_id']);
+            $table->index(
+                ['user_id', 'unlocked_at', 'unlocked_hardcore_at', 'achievement_id'],
+                'idx_user_date_achievement'
+            );
         });
     }
 
     public function down(): void
     {
         Schema::table('player_achievements', function (Blueprint $table) {
-            $table->dropIndex(['user_id', 'unlocked_at', 'unlocked_hardcore_at', 'achievement_id']);
+            $table->dropIndex('idx_user_date_achievement');
         });
     }
 };