--- a/database/migrations/platform/2023_10_30_000000_update_player_achievements_table.php
+++ b/database/migrations/platform/2023_10_30_000000_update_player_achievements_table.php
@@ -10,17 +10,27 @@ return new class() extends Migration {
     public function up(): void
     {
         Schema::table('player_achievements', function (Blueprint $table) {
-            $table->index(
-                ['user_id', 'unlocked_at', 'unlocked_hardcore_at', 'achievement_id'],
-                'player_achievements_user_date_achievement'
-            );
+            $sm = Schema::getConnection()->getDoctrineSchemaManager();
+            $indexesFound = $sm->listTableIndexes('player_achievements');
+
+            if (!array_key_exists('player_achievements_user_date_achievement', $indexesFound)) {
+                $table->index(
+                    ['user_id', 'unlocked_at', 'unlocked_hardcore_at', 'achievement_id'],
+                    'player_achievements_user_date_achievement'
+                );
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('player_achievements', function (Blueprint $table) {
-            $table->dropIndex('player_achievements_user_date_achievement');
+            $sm = Schema::getConnection()->getDoctrineSchemaManager();
+            $indexesFound = $sm->listTableIndexes('player_achievements');
+
+            if (array_key_exists('player_achievements_user_date_achievement', $indexesFound)) {
+                $table->dropIndex('player_achievements_user_date_achievement');
+            }
         });
     }
 };