`getAchievementsEarnedBetween()` is responsible for building the chart to show the user's points progress over the last ~2 weeks.

This is currently very slow for users with a lot of progress on the site, even though it is using denormalized data. After running `EXPLAIN` on the query, I identified the main driver behind this slowness is because the query has to scan potentially tens of thousands of rows on the `player_achievements` table due to a missing compound index.

**Benchmark Query**
```sql
SELECT
	COALESCE(pa.unlocked_hardcore_at, pa.unlocked_at) AS Date,
	SUM( IF(pa.unlocked_hardcore_at IS NOT NULL, ach.Points, 0)) AS HardcorePoints,
	SUM(ach.Points) AS SoftcorePoints
FROM
	player_achievements pa
	INNER JOIN Achievements AS ach ON ach.ID = pa.achievement_id
	INNER JOIN GameData AS gd ON gd.ID = ach.GameID
WHERE -- user = MaxMilyin
	pa.user_id = 16446 AND ach.Flags = 3 AND COALESCE(pa.unlocked_hardcore_at, pa.unlocked_at) BETWEEN '2023-10-16 21:52:14' AND '2023-10-30 21:52:14'
GROUP BY
	Date
ORDER BY
	Date ASC
LIMIT 0, 14;
```

**Before**
3.049s
2.748s
2.722s
2.773s
2.765s
Average: 2.81s

**After**
21ms
25ms
26ms
22ms
26ms
Average: 24ms